requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
discord.py==1.5.1
idna==2.10
isort==5.7.0
keyboard==0.13.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.4
pandas==1.1.5
Pygments==2.7.3
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
rich==9.5.1
shiboken2==5.15.2
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
yarl==1.5.1
